+++ src/components/Popover/example/index.js
@@ -1,0 +1,18 @@
+/* eslint-disable react/jsx-key */
+import React from 'react';
+import Popover from '../index';
+
+export default [
+  {
+    examples: [
+      <div className="position--relative height--100">
+        <Popover title="Lorem ipsum dolor sit amet">
+          <p>
+            Ipsa officiis bad-news minus earum a aperiam! Aperiam reiciendis vitae
+            nihil libero et, hic!
+          </p>
+        </Popover>
+      </div>,
+    ],
+  },
+];