+++ src/components/TabNav/Tab/tests/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Tab from 'components/TabNav/Tab';
+import Tab from '../index';
 import { shallow } from 'enzyme';
 
 describe('components/TabNav/Tab', () => {