+++ src/components/Table/tests/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Table from 'components/Table';
+import Table from '../index';
 import { shallow, render } from 'enzyme';
 
 describe('components/Table', () => {